fixed auth routing issue
